+++ dev-docs/CHANGELOG.md
@@ -9,16 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Language features
 
+- Optimized `Context().sender` to use `sender()` function for better gas efficiency: PR [#2427](https://github.com/tact-lang/tact/pull/2427)
 - [fix] Ternary operator with struct and null: PR [#2432](https://github.com/tact-lang/tact/pull/2432)
 
 ### Internal infrastructure
 
 - `internalExternalReceiversOutsideMethodsMap` have been reworked to ensure compatibility with explorers: PR [#2398](https://github.com/tact-lang/tact/pull/2398)
 
+### Docs
+
+- Fixed description of the `dump()` function, which does not support values of `StringBuilder` type: PR [#2463](https://github.com/tact-lang/tact/pull/2463)
+
 ### Release contributors
 
 - [Shvetc Andrei](https://github.com/Shvandre)
 - [Daniil Sedov](https://github.com/Gusarich)
+- [Novus Nota](https://github.com/novusnota)
 
 ## [1.6.4] - 2025-03-18